Completed controllers, specs, refactored task model and created .env example file
