+++ .env.example
@@ -1,0 +1,4 @@
+TWILIO_ACCOUNT_SID=<Your Twilio SID>
+TWILIO_AUTH_TOKEN=<Your Twilio Auth Token>
+TWILIO_NUMBER=<Your Twilio Account Phone Number>
+MY_NUMBER=<The Phone Number You Want To Send SMS Messages>
