+++ config/application.rb
@@ -11,6 +11,7 @@ require 'action_controller/railtie'
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
 Dotenv.load
+
 module Tracker
   class Application < Rails::Application
     # Settings in config/environments/* take precedence over those specified here.
