+++ spec/controllers/V1/tasks_controller_spec.rb
@@ -1,7 +1,7 @@
 require 'rails_helper'
-require 'webmock/rspec'
 
 describe V1::TasksController, :type => :controller do
+
   let!(:project) { create(:project) }
 
   describe "PUT update_task_state" do
@@ -24,24 +24,21 @@ describe V1::TasksController, :type => :controller do
       expect(task.state).to eql("in-progress")
     end
 
-    context "when a task is updated to complete" do
+    context "when a task is updated to done" do
       let!(:update_params) do { project_id: project.id, task_id: task.id, state: "done" } end
 
-        before do
-          stub_request(:post, "https://api.twilio.com/2010-04-01/Accounts/ACb44804bc4b710ac8dba26b98b0c60eae/Messages.json").
-            with(:headers => {'Accept'=>'application/json', 'Accept-Charset'=>'utf-8', 'Accept-Encoding'=>'gzip;q=1.0,deflate;q=0.6,identity;q=0.3', 'Authorization'=>'Basic QUNiNDQ4MDRiYzRiNzEwYWM4ZGJhMjZiOThiMGM2MGVhZTo2MmI1NTUwOTQwMzEwZjRmYzU3YmExOTg0NDcyYmE3MA==', 'Content-Type'=>'application/x-www-form-urlencoded', 'User-Agent'=>'twilio-ruby/4.11.1 (ruby/x86_64-darwin16 2.3.0-p0)'}).
-          to_return(:status => 200, :body => "", :headers => {})
-        #   stub_request(:post, "https://api.twilio.com/2010-04-01/Accounts/ACb44804bc4b710ac8dba26b98b0c60eae/Messages.json").
-        #  with(:body => {"Body"=>"Hooray! Task Complete!", "From"=>"+15412554232", "To"=>"+15624726279"},
-        #       :headers => {'Accept'=>'application/json', 'Accept-Charset'=>'utf-8', 'Accept-Encoding'=>'gzip;q=1.0,deflate;q=0.6,identity;q=0.3', 'Authorization'=>'Basic QUNiNDQ4MDRiYzRiNzEwYWM4ZGJhMjZiOThiMGM2MGVhZTo2MmI1NTUwOTQwMzEwZjRmYzU3YmExOTg0NDcyYmE3MA==', 'Content-Type'=>'application/x-www-form-urlencoded', 'User-Agent'=>'twilio-ruby/4.11.1 (ruby/x86_64-darwin16 2.3.0-p0)'}).
-        #  to_return(:status => 200, :body => "", :headers => {})
-        end
-
       it "sends an sms message when updated to done" do
-        put 'update_task_state', update_params        #
-        expect(a_request(:post, "https://api.twilio.com")).to have_been_made.times(1)
-        expect(task).to receive(:send_message_via_sms).with("Hooray! Task Complete!")
+        put 'update_task_state', update_params
+        body = JSON.parse(response.body)
+        expect(body).to include("completed")
       end
+
+      it "does not send an sms message if the state isn't 'done'" do
+        put 'update_task_state', update_params.merge(state: "in-progress")
+        body = JSON.parse(response.body)
+        expect(body).not_to include("completed")
+      end
+
     end
   end
