+++ spec/spec_helper.rb
@@ -14,9 +14,11 @@
 #
 # The `.rspec` file also contains a few flags that are not defaults but that
 # users commonly want.
-# WebMock.disable_net_connect!(allow_localhost: true)
+
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
+
+
   # rspec-expectations config goes here. You can use an alternate
   # assertion/expectation library such as wrong or the stdlib/minitest
   # assertions if you prefer.
